requirements.txt
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.11.4
six==1.16.0
tzdata==2024.1
